Extract cache to file and add tests
