+++ packages/cli-kit/src/environment/spin-cache.ts
@@ -1,0 +1,9 @@
+let cachedSpinFQDN: string
+
+export function getCachedSpinFqdn(): string | undefined {
+  return cachedSpinFQDN
+}
+
+export function setCachedSpinFqdn(fqdn: string) {
+  cachedSpinFQDN = fqdn
+}
