+++ packages/cli-kit/src/environment/spin.test.ts
@@ -1,17 +1,15 @@
 import {show, fqdn, isSpin, instance, workspace, namespace, host, SpinInstanceNotFound} from './spin.js'
-import {getCachedSpinFqdn} from './utilities.js'
+import {getCachedSpinFqdn, setCachedSpinFqdn} from './spin-cache.js'
 import {captureOutput} from '../system.js'
-import {describe, test, expect, vi, it, beforeAll} from 'vitest'
+import {describe, test, expect, vi, it} from 'vitest'
 
 vi.mock('../system')
+vi.mock('./spin-cache')
 
 const mockedCaptureOutput = vi.mocked(captureOutput)
 
 describe('fqdn', () => {
-  beforeAll(() => {
-    vi.mock('./utilities')
-  })
-  it('shows the latest when SPIN_INSTANCE is not present', async () => {
+  it('shows the latest when SPIN_INSTANCE is not present and there is no cached value', async () => {
     // Given
     const env = {}
     const showResponse = {fqdn: 'fqdn'}
@@ -24,9 +22,10 @@ describe('fqdn', () => {
     // Then
     expect(got).toEqual('fqdn')
     expect(mockedCaptureOutput).toHaveBeenCalledWith('spin', ['show', '--latest', '--json'], {env})
+    expect(setCachedSpinFqdn).toBeCalledWith('fqdn')
   })
 
-  it("doesn't show the latest when SPIN_INSTANCE is present", async () => {
+  it("doesn't show the latest when SPIN_INSTANCE is present and there is no cached value", async () => {
     // Given
     const env = {SPIN_INSTANCE: 'instance'}
     const showResponse = {fqdn: 'fqdn'}
@@ -39,6 +38,21 @@ describe('fqdn', () => {
     // Then
     expect(got).toEqual('fqdn')
     expect(mockedCaptureOutput).toHaveBeenCalledWith('spin', ['show', '--json'], {env})
+    expect(setCachedSpinFqdn).toBeCalledWith('fqdn')
+  })
+
+  it('return cached spin fqdn if valid', async () => {
+    // Given
+    const env = {}
+    vi.mocked(getCachedSpinFqdn).mockReturnValue('cachedFQDN')
+
+    // When
+    const got = await fqdn(env)
+
+    // Then
+    expect(got).toEqual('cachedFQDN')
+    expect(mockedCaptureOutput).not.toHaveBeenCalledWith('spin', ['show', '--json'], {env})
+    expect(setCachedSpinFqdn).not.toBeCalled()
   })
 })
