+++ packages/cli-kit/src/environment/utilities.ts
@@ -19,13 +19,3 @@ export function isSet(variable: string | undefined): boolean {
   }
   return true
 }
-
-let cachedSpinFQDN: string
-
-export function getCachedSpinFqdn(): string | undefined {
-  return cachedSpinFQDN
-}
-
-export function setCachedSpinFqdn(fqdn: string) {
-  cachedSpinFQDN = fqdn
-}
